refactor: backout unnecessary platform support, tidy ups
