+++ .github/workflows/test.yml
@@ -42,12 +42,16 @@ jobs:
       - name: go build
         run: |
           go build -v ./...
-          env CGO_ENABLED=0 GOOS=linux GOARCH=amd64 go build -v ./...
+
           # Compile without optimization to check potential stack overflow.
           # The option '-gcflags=all=-N -l' is often used at Visual Studio Code.
           # See also https://go.googlesource.com/vscode-go/+/HEAD/docs/debugging.md#launch and the issue hajimehoshi/ebiten#2120.
           go build "-gcflags=all=-N -l" -v ./...
 
+          # Check cross-compiling Linux binaries.
+          env GOOS=linux GOARCH=amd64 go build -v ./...
+          env GOOS=linux GOARCH=arm64 go build -v ./...
+          
           # Check cross-compiling WebAssembly.
           # Unfortunately it is difficult to test Oto on browsers since this requires user interactions.
           env GOOS=js GOARCH=wasm go build -v ./...
